Added settings for the setMaxIdleSwap and setMaxIdleBackup so the configuration can occur in the Manager node.
Also overrode the processExpires method and commented out the processExpires
call. This cause issues to due the eventually consistant model for DynamoDb.

Writes aren't always available immediately after the call completes, so there
are scenarios where the session will be written to the store and immediately
read back in to process the expired sessions. The session in the dynamodb
store was correct, but the session would be populated with the previous value
since it was immediately read back in.